# pkg/entrypoints/serve/rootapi/api.gen.go
// Package rootapi provides primitives to interact with the openapi HTTP API.
//
// Code generated by github.com/oapi-codegen/oapi-codegen/v2 version v2.8.0 DO NOT EDIT.
package rootapi

import (
	"fmt"
	"net/http"

	"github.com/labstack/echo/v5"
	"github.com/oapi-codegen/runtime"
)

// ServerInterface represents all server handlers.
type ServerInterface interface {
	// GetNarPathNar Get the NAR file for a given hash
	// (GET /nar/{path}.nar)
	GetNarPathNar(ctx *echo.Context, path string) error
	// GetNixCacheInfo Get the binary cache configuration information
	// (GET /nix-cache-info)
	GetNixCacheInfo(ctx *echo.Context) error
	// GetNixhashNarinfo Get the NAR info description of a nix hash
	// (GET /{nixhash}.narinfo)
	GetNixhashNarinfo(ctx *echo.Context, nixhash string) error
}

// ServerInterfaceWrapper converts echo contexts to parameters.
type ServerInterfaceWrapper struct {
	Handler ServerInterface
}

// GetNarPathNar converts echo context to params.
func (w *ServerInterfaceWrapper) GetNarPathNar(ctx *echo.Context) error {
	var err error
	// ------------- Path parameter "path" -------------
	var path string

	err = runtime.BindStyledParameterWithOptions("simple", "path", ctx.Param("path"), &path, runtime.BindStyledParameterOptions{ParamLocation: runtime.ParamLocationPath, Explode: false, Required: true, Type: "string", Format: "", ValueIsUnescaped: ctx.Request().URL.RawPath == ""})
	if err != nil {
		return echo.NewHTTPError(http.StatusBadRequest, fmt.Sprintf("Invalid format for parameter path: %s", err))
	}

	// Invoke the callback with all the unmarshaled arguments
	err = w.Handler.GetNarPathNar(ctx, path)
	return err
}

// GetNixCacheInfo converts echo context to params.
func (w *ServerInterfaceWrapper) GetNixCacheInfo(ctx *echo.Context) error {
	var err error

	// Invoke the callback with all the unmarshaled arguments
	err = w.Handler.GetNixCacheInfo(ctx)
	return err
}

// GetNixhashNarinfo converts echo context to params.
func (w *ServerInterfaceWrapper) GetNixhashNarinfo(ctx *echo.Context) error {
	var err error
	// ------------- Path parameter "nixhash" -------------
	var nixhash string

	err = runtime.BindStyledParameterWithOptions("simple", "nixhash", ctx.Param("nixhash"), &nixhash, runtime.BindStyledParameterOptions{ParamLocation: runtime.ParamLocationPath, Explode: false, Required: true, Type: "string", Format: "", ValueIsUnescaped: ctx.Request().URL.RawPath == ""})
	if err != nil {
		return echo.NewHTTPError(http.StatusBadRequest, fmt.Sprintf("Invalid format for parameter nixhash: %s", err))
	}

	// Invoke the callback with all the unmarshaled arguments
	err = w.Handler.GetNixhashNarinfo(ctx, nixhash)
	return err
}

// This is a simple interface which specifies echo.Route addition functions which
// are present on both echo.Echo and echo.Group, since we want to allow using
// either of them for path registration
type EchoRouter interface {
	CONNECT(path string, h echo.HandlerFunc, m ...echo.MiddlewareFunc) echo.RouteInfo
	DELETE(path string, h echo.HandlerFunc, m ...echo.MiddlewareFunc) echo.RouteInfo
	GET(path string, h echo.HandlerFunc, m ...echo.MiddlewareFunc) echo.RouteInfo
	HEAD(path string, h echo.HandlerFunc, m ...echo.MiddlewareFunc) echo.RouteInfo
	OPTIONS(path string, h echo.HandlerFunc, m ...echo.MiddlewareFunc) echo.RouteInfo
	PATCH(path string, h echo.HandlerFunc, m ...echo.MiddlewareFunc) echo.RouteInfo
	POST(path string, h echo.HandlerFunc, m ...echo.MiddlewareFunc) echo.RouteInfo
	PUT(path string, h echo.HandlerFunc, m ...echo.MiddlewareFunc) echo.RouteInfo
	TRACE(path string, h echo.HandlerFunc, m ...echo.MiddlewareFunc) echo.RouteInfo
}

// RegisterHandlersOptions configures RegisterHandlersWithOptions.
type RegisterHandlersOptions struct {
	// BaseURL is prepended to every registered path so the API can be served
	// under a prefix.
	BaseURL string
	// OperationMiddlewares lets the caller attach per-operation middleware at
	// registration time. The map key is the OpenAPI `operationId` value as it
	// appears in the spec (the raw, un-normalized form). Operations that have
	// no entry are registered with no extra middleware. A nil map disables
	// per-operation middleware entirely.
	OperationMiddlewares map[string][]echo.MiddlewareFunc
}

// RegisterHandlers adds each server route to the EchoRouter.
func RegisterHandlers(router EchoRouter, si ServerInterface) {
	RegisterHandlersWithOptions(router, si, RegisterHandlersOptions{})
}

// RegisterHandlersWithBaseURL registers handlers and prepends BaseURL to the
// paths so the API can be served under a prefix.
func RegisterHandlersWithBaseURL(router EchoRouter, si ServerInterface, baseURL string) {
	RegisterHandlersWithOptions(router, si, RegisterHandlersOptions{BaseURL: baseURL})
}

// RegisterHandlersWithOptions registers handlers using the supplied options,
// including any per-operation middleware.
func RegisterHandlersWithOptions(router EchoRouter, si ServerInterface, options RegisterHandlersOptions) {

	wrapper := ServerInterfaceWrapper{
		Handler: si,
	}

	router.GET(options.BaseURL+"/nix-cache-info", wrapper.GetNixCacheInfo, options.OperationMiddlewares["GetNixCacheInfo"]...)
	router.GET(options.BaseURL+"/:nixhash.narinfo", wrapper.GetNixhashNarinfo, options.OperationMiddlewares["GetNixhashNarinfo"]...)
	router.GET(options.BaseURL+"/nar/:path.nar", wrapper.GetNarPathNar, options.OperationMiddlewares["GetNarPathNar"]...)

}